# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Mobile Shop Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /expenses/{id}:
    delete:
      operationId: deleteExpense
      tags: [expenses]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── EMPLOYEES ───────────────────────────────────────────────
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      responses:
        "200":
          description: Employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [employees]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}:
    patch:
      operationId: updateEmployee
      tags: [employees]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    delete:
      operationId: deleteEmployee
      tags: [employees]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── INSTALLMENTS ────────────────────────────────────────────
  /installments:
    get:
      operationId: listInstallments
      tags: [installments]
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Installment plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InstallmentPlan"
    post:
      operationId: createInstallment
      tags: [installments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InstallmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstallmentPlan"
  /installments/{id}:
    get:
      operationId: getInstallment
      tags: [installments]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Installment plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstallmentPlan"
    patch:
      operationId: updateInstallment
      tags: [installments]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InstallmentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstallmentPlan"

  # ─── WALLETS ─────────────────────────────────────────────────
...
  /fawry/balance:
    get:
      operationId: getFawryBalance
      tags: [fawry]
      summary: Get current Fawry balance summary
      responses:
        "200":
          description: Fawry balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FawryBalance"
    post:
      operationId: addFawryBalance
      tags: [fawry]
      summary: Record received balance from agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FawryBalanceInput"
      responses:
        "201":
          description: Recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FawryBalance"
  /fawry/transactions:
    get:
      operationId: listFawryTransactions
      tags: [fawry]
      responses:
        "200":
          description: Fawry transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FawryTransaction"
    post:
      operationId: createFawryTransaction
      tags: [fawry]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FawryTransactionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FawryTransaction"

  # ─── REPORTS ─────────────────────────────────────────────────
...
components:
  schemas:
...
    Expense:
      type: object
      required: [id, category, amount, date, createdAt]
      properties:
        id:
          type: integer
        category:
          type: string
        amount:
          type: number
        date:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    ExpenseInput:
      type: object
      required: [category, amount, date]
      properties:
        category:
          type: string
        amount:
          type: number
        date:
          type: string
        notes:
          type: string

    # ─── EMPLOYEES ───────────────────────────────────────────────
    Employee:
      type: object
      required: [id, name, phone, role, salary]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        role:
          type: string
        salary:
          type: number
        advances:
          type: number
        deductions:
          type: number
        createdAt:
          type: string
    EmployeeInput:
      type: object
      required: [name, phone, role, salary]
      properties:
        name:
          type: string
        phone:
          type: string
        role:
          type: string
        salary:
          type: number
    EmployeeUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        role:
          type: string
        salary:
          type: number
        advances:
          type: number
        deductions:
          type: number

    # ─── INSTALLMENTS ────────────────────────────────────────────
    InstallmentPlan:
      type: object
      required: [id, customerId, customerName, customerPhone, deviceName, totalAmount, downPayment, installmentAmount, totalInstallments, paidInstallments, remainingAmount, nextDueDate, status, createdAt]
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: string
        customerPhone:
          type: string
        deviceName:
          type: string
        totalAmount:
          type: number
        downPayment:
          type: number
        installmentAmount:
          type: number
        totalInstallments:
          type: integer
        paidInstallments:
          type: integer
        remainingAmount:
          type: number
        nextDueDate:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, completed, overdue]
        createdAt:
          type: string
    InstallmentInput:
      type: object
      required: [customerId, deviceName, totalAmount, downPayment, installmentAmount, totalInstallments, startDate]
      properties:
        customerId:
          type: integer
        deviceName:
          type: string
        totalAmount:
          type: number
        downPayment:
          type: number
        installmentAmount:
          type: number
        totalInstallments:
          type: integer
        startDate:
          type: string
    InstallmentUpdate:
      type: object
      properties:
        paidInstallments:
          type: integer
        status:
          type: string
          enum: [active, completed, overdue]

    # ─── WALLETS ─────────────────────────────────────────────────
...
    FawryBalance:
      type: object
      required: [received, used, remaining, todayProfit]
      properties:
        received:
          type: number
        used:
          type: number
        remaining:
          type: number
        todayProfit:
          type: number
    FawryBalanceInput:
      type: object
      required: [amount]
      properties:
        amount:
          type: number
        notes:
          type: string
    FawryTransaction:
      type: object
      required: [id, serviceType, amount, profit, createdAt]
      properties:
        id:
          type: integer
        serviceType:
          type: string
          enum: [recharge, scratch_card, recharge_card, electricity, water, gas, other]
        amount:
          type: number
        profit:
          type: number
        customerPhone:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    FawryTransactionInput:
      type: object
      required: [serviceType, amount, profit]
      properties:
        serviceType:
          type: string
          enum: [recharge, scratch_card, recharge_card, electricity, water, gas, other]
        amount:
          type: number
        profit:
          type: number
        customerPhone:
          type: string
        notes:
          type: string

    # ─── REPORTS ─────────────────────────────────────────────────
...